Build and reinstate paths to enable redirects
