+++ docs/developers/guides/run-a-node/besu.mdx
@@ -1,6 +1,7 @@
 ---
 title: Besu
 description: Install the Besu client to run a Linea node.
+image: /img/socialCards/besu.jpg
 ---
 
 import Tabs from "@theme/Tabs";
