+++ docs/developers/guides/run-a-node/bootnodes.mdx
@@ -1,7 +1,7 @@
 ---
 title: Bootnodes
 description: Bootnodes available for Linea Mainnet
-image: /img/socialCards/edge-nodes.jpg
+image: /img/socialCards/bootnodes.jpg
 ---
 
 The following bootnodes enable your node to find a peer node when initializing. To
