+++ docs/developers/guides/run-a-node/erigon.mdx
@@ -1,6 +1,7 @@
 ---
 title: Erigon
 description: Install the Erigon client to run a Linea node.
+image: /img/socialCards/erigon.jpg
 ---
 
 import Tabs from "@theme/Tabs";
