+++ docs/developers/guides/run-a-node/geth.mdx
@@ -1,6 +1,7 @@
 ---
 title: Geth
 description: Install the Geth client to run a Linea node.
+image: /img/socialCards/geth.jpg
 ---
 
 import Tabs from "@theme/Tabs";
