+++ docs/developers/guides/run-a-node/linea-besu.mdx
@@ -1,6 +1,7 @@
 ---
 title: Linea Besu
 description: Install the Besu client to run a Linea node.
+image: /img/socialCards/linea-besu.jpg
 ---
 
 import Tabs from "@theme/Tabs";
