+++ docs/users/index.mdx
@@ -1,0 +1,22 @@
+---
+title: About Linea
+sidebar_position: 2
+image: /img/socialCards/about-linea.jpg
+---
+
+Linea is an Ethereum layer 2 (L2) network that provides access to innovative Web3
+projects at significantly lower costs compared to Ethereum. It's compatible with the
+Ethereum Virtual Machine (EVM), allowing users to deploy existing Ethereum applications and
+develop new ones that would be too costly on Mainnet. It also enables the community to
+use those dapps, **at a fraction of the cost**, and at **multiples the speed** of Mainnet.
+
+Linea delivers unrivaled network security with its cryptographically secure zk-rollup infrastructure,
+comprehensive security partnerships, and security-oriented ecosystem culture. View the
+[Linea architecture](../architecture/index.mdx) for more information.
+
+You can get started by [funding your Linea account](./move-funds/index.mdx) using the
+fiat onramp in MetaMask Portfolio or by bridging over funds.
+
+:::tip
+If you run into a problem, view the [Linea Support page](https://support.linea.build/hc/en-us) and let us know.
+:::
