+++ docs/users/linea-voyage/index.mdx
@@ -1,0 +1,8 @@
+---
+title: Linea Voyage
+image: /img/socialCards/linea-voyage.jpg
+---
+
+import DocCardList from '@theme/DocCardList';
+
+<DocCardList />
