+++ docs/users/linea-voyage/linea-surge/index.mdx
@@ -1,0 +1,21 @@
+---
+title: 'Linea Voyage: The Surge'
+image: /img/socialCards/linea-voyage-the-surge.jpg
+---
+
+Linea Surge is an LXP-driven program that enables people to collect LXP-L in return for having 
+assets on Linea and deploying those assets into protocols.
+
+The goal for The Surge is to allow users that provide liquidity on the network to collect LXP-L. 
+LXP-L is similar to the system we have in place today for LXP, but is collected by contributing 
+liquidity to Linea. 
+
+A core focus is on the expansion of available assets through new applications 
+and DeFi use cases built on Linea. Additional liquidity brought by new applications sustainably 
+supports additional volume, in turn impacting on the sustainability of the liquidity through 
+increased fees for liquidity providers and increasing the percentage of ecosystem trading 
+activity compared to other venues. This constitutes something of a flywheel network effect for 
+the ecosystem.
+
+The following sections provide further detail. In the overview page, we will explore the program,
+before explaining the math behind it in _Point incentive design_.
