+++ docs/users/linea-voyage/linea-surge/linea-surge-overview.mdx
@@ -1,0 +1,65 @@
+---
+title: Overview
+sidebar_position: 1
+image: /img/socialCards/overview.jpg
+---
+
+Once it starts, the program will last for 6 months, or until the network TVL reaches $3B, whichever 
+happens first. The Surge is structured into six Volts (epochs), each one month long. Why Volts? 
+Because if we reach the target we will let the epoch run till the end of the Volt. And we can play 
+with themes: every Volt will have narratives and featured partners, and provide educational 
+content — each Volt a distinct experience.
+
+## Early adopters
+
+The points system will start when the program goes live but we already have users and early 
+adopters who have provided liquidity on the network. It's important to acknowledge what 
+has been done so far. For this reason, the formula that defines points has an early adopter 
+modifier that provides more points to users that started contributing earlier than others.
+
+## Dashboard
+
+After go-live, OpenBlock Labs will provide a dashboard where anyone can verify the number 
+of LXP-L collected within The Surge. This will feature all necessary information for a user to 
+observe their activity and points received.
+
+## Sybil
+
+While in previous campaigns we tied the minting of LXP to addresses that hold Proof Of Humanity 
+to specifically protect the LXP points from bot farms and reduce sybils, in this case there's 
+no such need. Providing a unit of points per $ locked is an action that does not incentivize bot 
+activity: a user will get the same amount of points regardless of whether they concentrate all 
+the liquidity in a single wallet or split it among thousands of addresses. This makes the 
+experience even easier and more enjoyable for users who bridge liquidity to Linea.
+
+## Security and UX
+
+Dapps that participate in the program will have a great opportunity to grow, but our main 
+priorities at Linea are security and UX. We can't afford to assign LXP-L points to dapps that 
+put users' funds at risk, so this is the right time to raise the bar for security at an ecosystem 
+level. Every dapp that is on the ecosystem portal today is audited by one of the Linea security 
+partners, and in order to be eligible for inclusion in the program, a dapp must satisfy the 
+following criteria:
+
+- Meet UX requirements from an established framework
+- Satisfy at least one of the two following conditions:
+    - Has been live for at least one year on any other EVM chain with a TVL peak of more than $50M
+    - Provide an audit report associated with the version of the contract that is deployed on 
+    Linea and has implemented at least one real-time exploit detection solution provided by 
+    Linea Partners
+
+## Relationship between LXP and LXP-L
+
+The goal for The Surge is to attract more liquidity by tracking the contribution of each user, 
+leveraging a similar system that we have in place today: LXP-L will be a different token with 
+the same non-transferable characteristic of LXP. Both of these points systems recognize involvement 
+from two distinctly different user bases while providing greater transparency for each group.
+
+This approach ensures a balanced alignment between LXP and LXP-L. Considering the complexity 
+of the Surge points model, it is challenging to establish precise alignment for these two types 
+of points. Implementing parallel points systems allows us to assess the alignment retrospectively
+and make necessary adjustments to maintain fairness between users who contribute liquidity 
+and those who contribute through activity.
+
+Moreover, we will continue to provide opportunities for users to contribute to the growth of 
+the ecosystem by performing activity-based activations on Linea dapps.
