+++ docs/users/linea-voyage/lxp/index.mdx
@@ -1,0 +1,149 @@
+---
+title: Linea Voyage LXP
+sidebar_position: 5
+image: /img/socialCards/linea-voyage-lxp.jpg
+---
+
+## What is Linea Voyage XP?
+
+Similar to video and role-playing games, wherein your character can gain Experience Points (XP) as
+you go through the game, Linea Voyage LXP represents your journey and experience in the Linea
+ecosystem.
+The Voyage LXP tokens will be a custom
+[ERC-20 token](https://ethereum.org/en/developers/docs/standards/tokens/erc-20/), and will be
+visible by default in your MetaMask wallet.
+
+**Voyage LXP tokens are non-transferable, soulbound tokens that are distributed to recognize the
+community's extraordinary contribution toward the growth of the Linea ecosystem. They do not have
+any monetary value!**
+
+Apart from being a measure of contribution to building Linea into a robust and secure L2 network,
+Voyage LXP will also make owners eligible for benefits such as receiving official community roles,
+exclusive Linea swag, and more!
+
+To learn more about the Linea Voyage program, click [here](https://support.linea.build/linea-voyage).
+
+:::note
+Voyage LXP are non-transferable, and cannot be bridged to other networks. You cannot buy, sell, or
+swap them, nor can you accumulate them by transferring LXP from another account.
+
+However, you can use multiple accounts to earn LXP; more on that below.
+:::
+
+## Contract address
+
+| Voyage LXP contract address |
+| --- |
+| [0xd83af4fbD77f3AB65C3B1Dc4B38D7e67AEcf599A](https://lineascan.build/address/0xd83af4fbD77f3AB65C3B1Dc4B38D7e67AEcf599A) |
+
+## Account binding
+
+<div class="center-container">
+    <div class="video-container">
+      <iframe
+        class="video-iframe"
+        src="https://www.youtube.com/embed/qH5xRiXSMSg?si=k3TO5LpHUWaR-TbU"
+        title="Linea Voyage LXP: Account Binding Guide"
+        allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
+        allowFullScreen></iframe>
+    </div>
+</div>
+
+Account binding allows you to aggregate your Voyage LXP tokens to a single primary account.
+If you're navigating the Linea ecosystem with multiple accounts, choose one as your main hub to
+effortlessly accumulate all the Voyage LXP you've received.
+This ensures that every adventure is accurately recorded across your interconnected accounts,
+providing a seamless and comprehensive representation of your Linea experience.
+
+**Account binding is not retroactive. If you bind your account address after Voyage XP is received,
+then those won't be linked to the primary account!**
+
+:::caution
+
+**Account binding is permanent and irreversible, so make sure you are choosing the correct account address when going through the process!**
+
+- Using account binding is only necessary for those who want to bind two or more accounts.
+
+- To receive LXP, the Primary account set must pass Proof of Humanity. Accounts set as secondary do not need to pass PoH.
+
+- All LXP will be sent to your Primary account when using Account Binding.
+
+- You will receive LXP once per task. The exception to this is the Linea Voyage Testnet NFT, holders will receive LXP for each NFT owned.
+:::
+
+To start the Account binding process, follow the steps below:
+
+1. Go [here](https://linea.build/activations)
+2. Connect your wallet in the top-right corner of the page
+
+<div class="center-container">
+  <div class="img-large">
+    <img
+      src="/img/article_images/Use_Linea/Linea_voyage_xp/Linea_Voyage_connect_wallet.png"
+      alt="Linea Voyage LXP connect wallet"
+    />
+  </div>
+</div>
+
+3. Click on the address of your currently selected account to open a dropdown menu. Here, click
+"Manage Wallet(s)"
+
+<div class="center-container">
+  <div class="img-large">
+    <img
+      src="/img/article_images/Use_Linea/Linea_voyage_xp/Linea_Voyage_manage_wallets.png"
+      alt="Linea Voyage LXP manage wallets"
+    />
+  </div>
+</div>
+
+4. Sign the signature request
+
+<div class="center-container">
+  <div class="img-small">
+    <img
+      src="/img/article_images/Use_Linea/Linea_voyage_xp/Linea_Voyage_signature_request.png"
+      alt="MetaMask Linea signature request"
+    />
+  </div>
+</div>
+
+5. Select "Set as Primary Account"
+
+<div class="center-container">
+  <div class="img-small">
+    <img
+      src="/img/article_images/Use_Linea/Linea_voyage_xp/Linea_Voyage_set_primary_account.png"
+      alt="Linea voyage set as primary account"
+    />
+  </div>
+</div>
+
+6. Select "Add secondary Account"
+
+<div class="center-container">
+  <div class="img-small">
+    <img
+      src="/img/article_images/Use_Linea/Linea_voyage_xp/Linea_Voyage_set_secondary_account.png"
+      alt="Linea voyage add secondary account"
+    />
+  </div>
+</div>
+
+7. Follow the prompts to connect the account you want to bind and sign the signature request.
+8. Read through the disclaimer, check the box, and hit confirm to finish the binding process
+
+<div class="center-container">
+  <div class="img-small">
+    <img
+      src="/img/article_images/Use_Linea/Linea_voyage_xp/Linea_Voyage_confirm_accounts.png"
+      alt="Linea voyage confirm secondary account"
+    />
+  </div>
+</div>
+
+:::info[What if I want to bind an account from a different Secret Recovery Phrase/seed phrase?]
+
+Refer to the [support guide](https://support.linea.build/linea-voyage/lxp-binding-an-account-from-a-different-secret-recovery-phrase-).
+
+:::
