+++ docs/users/move-funds/bridge.mdx
@@ -1,0 +1,42 @@
+---
+title: Bridge your funds
+image: /img/socialCards/bridge-your-funds.jpg
+---
+
+The [official Linea bridge](https://bridge.linea.build/) provides multiple options to bridge
+your funds to Linea.
+
+You can select the option most appropriate to you:
+
+- **MetaMask bridge for users** - The bridge feature in MetaMask Portfolio, which we recommend as
+    the simplest option for most users.
+- **Third-party bridges for users** - Use alternative third-party bridges. 
+- **Linea Native Bridge** - Best suited for tech operators, dapp teams, and liquidity 
+providers. For more guidance on the native bridge, see [here](../../developers/guides/bridge/how-to-bridge-eth.mdx).
+
+:::info
+The Linea Native Bridge is a slower option that should not be used for time-sensitive transfers.
+Additionally, you will need to manually claim ERC-20 tokens that are transferred across using this bridge.
+:::
+
+## Use the MetaMask Portfolio bridge
+
+The [MetaMask Portfolio's bridge](https://portfolio.metamask.io/bridge) surveys bridges across Linea to
+source its routes and rates and provides you with the best price.
+
+MetaMask Portfolio's bridge also allows you to bridge and swap tokens as part of the same transaction.
+
+[View the MetaMask Portfolio's bridge documentation](https://support.metamask.io/metamask-portfolio/bridge/)
+for more information. 
+
+## Third-party bridges
+
+You can find third-party bridges on the [ecosystem portal](https://linea.build/apps). Use the
+filter to find the available third-party bridges.
+
+:::caution
+
+These are permissionless bridges, and, you should do your own research before interacting
+with dapps in order to keep your funds safe.
+
+:::
