+++ docs/users/move-funds/fund.mdx
@@ -1,0 +1,78 @@
+---
+title: Fund your accounts
+description: Get tokens in your accounts
+sidebar_position: 2
+image: /img/socialCards/fund-your-accounts.jpg
+---
+
+Linea uses ETH for network transaction fees. You can fund your account on either [Linea Mainnet](#linea-mainnet) for
+real transactions or [Linea Sepolia](#linea-sepolia) for testing purposes.
+
+## Linea Mainnet
+
+You can buy tokens directly on Linea using [MetaMask Portfolio](https://portfolio.metamask.io/).
+Alternatively, if you already have tokens on another network, you can [bridge them to Linea](./bridge.mdx).
+For example you can bridge ETH on the Ethereum network over to the Linea network.
+
+:::info
+The [MetaMask Portfolio](https://portfolio.metamask.io/) dapp allows you to buy, sell, swap, stake, bridge, and
+manage your cryptoassets, among other features.
+
+Linea purchases in MetaMask Portfolio are available worldwide.
+::: 
+
+<div class="center-container">
+  <div class="video-container">
+    <iframe
+      class="video-iframe"
+      src="//www.youtube.com/embed/BOPDw348-zg"
+      title="TITLE"
+      allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
+      allowFullScreen></iframe>
+  </div>
+</div>
+
+### Buy tokens on Linea Mainnet
+
+To buy tokens using MetaMask Portfolio:
+
+1. Access [MetaMask Portfolio](https://portfolio.metamask.io), and connect the account that will
+    receive your purchased tokens.
+1. Select the option to buy tokens:
+    - On a mobile device, select **Buy** from the hamburger icon menu located in the top left 
+        of the screen.
+    - On a desktop, select **Buy** on screen.
+
+        :::info
+        If you've never used the **Buy** option previously, you'll first need to 
+        select your region because payment methods and provider availability vary based on your bank's location.
+        :::
+
+1. Set the payment method, if you've bought tokens before on your current device, your 
+    previous choice is loaded by default.
+1. In the **Buy** menu, select the token, and locate and select **Linea** from the networks list.
+1. In the list of tokens available to buy on Linea, select the one you want to buy.
+1. On the main screen, input the amount you wish to spend.
+1. View your quotes:
+    - On a desktop browser, one or more quotes appear on the right.
+    - On mobile devices, select **Get Quotes**.
+1. Pick a quote by selecting it.
+
+Selecting a quote redirects you to the provider's site to complete the transaction. You may need 
+to complete KYC ("Know Your Customer"; essentially, identifying yourself) with the provider.
+
+After you've completed the steps with the provider, you'll be routed back to MetaMask Portfolio. 
+Timescales for receiving your funds vary based on the provider and token route.
+
+## Linea Sepolia
+
+You can use one of the following Linea Sepolia faucets to drip testnet ETH directly to your
+Linea Sepolia address:
+
+
+- [Covalent](https://www.covalenthq.com/faucet/)
+- [Infura](https://www.infura.io/faucet/linea)
+- [GetBlock](https://getblock.io/faucet/linea-sepolia/)
+- [HackQuest](https://www.hackquest.io/en/faucets/59141) (and check out their [Linea Learning Track](https://www.hackquest.io/en/learning-track/9be129e7-575b-49bd-a64e-1bbe32427ace))
+
+Alternatively, [use the Linea Native Bridge to bridge tokens between Sepolia and Linea Sepolia](../../developers/guides/bridge/how-to-bridge-eth.mdx).
