+++ docs/users/move-funds/index.mdx
@@ -1,0 +1,8 @@
+---
+title: Move funds to Linea
+image: /img/socialCards/move-funds-to-linea.jpg
+---
+
+import DocCardList from '@theme/DocCardList';
+
+<DocCardList />
