+++ docs/users/move-funds/set-up-your-wallet.mdx
@@ -1,0 +1,70 @@
+---
+title: Set up your wallet
+description: Set up your wallet to start using Linea
+sidebar_position: 1
+image: /img/socialCards/set-up-your-wallet.jpg
+---
+
+Use a wallet such as MetaMask to manage your digital assets and interact with dapps on the Linea network.
+
+:::info
+
+Linea is designed to be completely compatible with Ethereum and works with 
+MetaMask out of the box.
+
+[View the MetaMask support site](https://support.metamask.io/getting-started/getting-started-with-metamask/) 
+if you're not familiar with the MetaMask wallet.
+
+:::
+
+## Install MetaMask
+
+The MetaMask browser extension and mobile app provide built-in connections to Linea.
+
+You can [install MetaMask as a browser extension or mobile app](https://metamask.io/download/).
+
+## Access Linea on MetaMask
+
+### Linea Mainnet
+
+Linea Mainnet is included in the available networks in the MetaMask browser extension and mobile 
+app by default.
+
+Select **Linea Mainnet** from the network dropdown to connect to the network.
+[View the MetaMask instructions](https://support.metamask.io/networks-and-sidechains/managing-networks/how-to-connect-to-linea-on-metamask/)
+for more information.
+
+### Linea Sepolia
+
+To connect to Linea Sepolia testnet, select **Show test networks**, as indicated in the image 
+below, and then select **Linea Sepolia**.
+
+<div class="center-container">
+  <div class="img-small">
+    <img
+      src="/img/article_images/Use_Linea/Set_up_your_wallet/Linea_find_Linea_sepolia2.gif"
+      alt="Show Linea Sepolia network MetaMask extension"
+    />
+  </div>
+</div>
+
+## Next steps
+ 
+View [the ecosystem portal](https://linea.build/apps) to see list of Web3 dapps available on Linea.
+
+Regardless of what you're doing on Linea, you'll likely need some tokens to transact on the network. 
+To do this, you can [fund your Linea account](./fund.mdx).
+
+## Troubleshooting tips
+
+If you're interacting with a dapp on Linea, try that dapp's documentation and resources first. 
+If you can't find your answer there, take a look at the [Linea Support Help Center](https://support.linea.build), 
+and search on the [Linea Community Forums](https://community.linea.build) or [Discord](https://discord.com/invite/linea) 
+for related issues. If it seems like a MetaMask issue at that point, jump over to 
+[MetaMask's Help Center](https://support.metamask.io), where you can explore their forums, and 
+start a conversation with support.
+
+## Other wallets
+
+Linea is compatible with any EVM-compatible wallet. If you want to add the Linea network, then
+[view the relevant network information](../../developers/quickstart/info-contracts.mdx).
