--- scripts/redirects.js
@@ -1,9 +1,0 @@
-const redirects = {
-    '/users/linea-voyage/linea-surge/linea-surge-overview': 'https://support.linea.build/linea-voyage/linea-surge/linea-surge-overview',
-    '/users/linea-voyage/linea-surge/linea-surge-model': 'https://support.linea.build/linea-voyage/linea-surge/linea-surge-model',
-    '/users/linea-voyage/linea-surge': 'https://support.linea.build/linea-voyage/linea-surge',
-    '/users/linea-voyage/lxp': 'https://support.linea.build/linea-voyage/lxp',
-    // Add more paths as needed
-  };
-  
-  export default redirects;
